backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0

# Redis
redis==5.2.0
hiredis==3.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.1

# Pydantic
pydantic==2.10.3
pydantic-settings==2.6.1

# OpenAI
openai==1.57.4

# Image processing
Pillow==11.0.0
imagehash==4.3.1

# File handling
aiofiles==24.1.0
python-magic==0.4.27

# S3/MinIO (optional)
boto3==1.35.80
aioboto3==13.3.0

# Background tasks - Celery for production
celery[redis]==5.4.0
flower==2.0.1  # Celery monitoring UI (optional)

# Monitoring
prometheus-client==0.21.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Development
black==24.10.0
isort==5.13.2
mypy==1.14.0


